requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.89.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gradio==3.16.2
gTTS==2.3.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
idna==3.4
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==9.0.0
multidict==6.0.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai-whisper==20230124
orjson==3.8.5
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyChatGPT==0.4.3.3
pycryptodome==3.16.0
pydantic==1.10.4
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
PyVirtualDisplay==3.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.22.0
tokenizers==0.13.2
toolz==0.12.0
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uc-micro-py==1.0.1
undetected-chromedriver==3.2.1
urllib3==1.26.14
uvicorn==0.20.0
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
